requirements.txt
yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
praw==7.7.1
feedparser==6.0.11
nltk==3.9.1
schedule==1.2.2
PyYAML==6.0.2
matplotlib==3.9.0
alpaca-py==0.43.2
beautifulsoup4==4.12.3
